+++ apps/integration-tests/src/test/kotlin/io/stereov/singularity/ratelimit/RateLimitIpFilterTest.kt
@@ -1,7 +1,7 @@
 package io.stereov.singularity.ratelimit
 
 import io.stereov.singularity.ratelimit.properties.RateLimitProperties
-import io.stereov.singularity.secrets.core.properties.KeyManagerImplementation
+import io.stereov.singularity.secrets.core.properties.SecretStoreImplementation
 import io.stereov.singularity.test.BaseSpringBootTest
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -34,7 +34,7 @@ class RateLimitIpFilterTest : BaseSpringBootTest() {
         fun properties(registry: DynamicPropertyRegistry) {
             registry.add("singularity.app.support-email") { "support@example.com" }
             registry.add("singularity.app.create-root-user") { "false" }
-            registry.add("singularity.secrets.key-manager") { KeyManagerImplementation.Mock }
+            registry.add("singularity.secrets.store") { SecretStoreImplementation.Mock }
             registry.add("singularity.secrets.bitwarden.api-url") { "https//api.bitwarden.com" }
             registry.add("singularity.secrets.bitwarden.identity-url") { "https//identity.bitwarden.com" }
             registry.add("singularity.secrets.bitwarden.organization-id") { UUID.randomUUID() }