+++ apps/integration-tests/src/test/kotlin/io/stereov/singularity/test/config/MockConfig.kt
@@ -12,7 +12,7 @@ class MockConfig {
     fun googleAuthenticator(): GoogleAuthenticator = mockk(relaxed = true)
 
     @Bean
-    fun keyManager(): MockKeyManager = MockKeyManager()
+    fun keyManager(): MockSecretStore = MockSecretStore()
 
     @Bean
     fun fileStorage(): FileStorage = mockk(relaxed = true)