+++ libs/core/src/main/kotlin/io/stereov/singularity/encryption/config/EncryptionConfiguration.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import io.stereov.singularity.encryption.service.EncryptionSecretService
 import io.stereov.singularity.encryption.service.EncryptionService
 import io.stereov.singularity.global.properties.AppProperties
-import io.stereov.singularity.secrets.core.component.KeyManager
+import io.stereov.singularity.secrets.core.component.SecretStore
 import io.stereov.singularity.secrets.bitwarden.config.BitwardenSecretsConfiguration
 import io.stereov.singularity.secrets.core.config.SecretsConfiguration
 import org.springframework.boot.autoconfigure.AutoConfiguration
@@ -23,13 +23,13 @@ class EncryptionConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    fun encryptionSecretService(keyManager: KeyManager, appProperties: AppProperties): EncryptionSecretService {
-        return EncryptionSecretService(keyManager, appProperties)
+    fun encryptionSecretService(secretStore: SecretStore, appProperties: AppProperties): EncryptionSecretService {
+        return EncryptionSecretService(secretStore, appProperties)
     }
 
     @Bean
     @ConditionalOnMissingBean
-    fun encryptionService(encryptionSecretService: EncryptionSecretService, keyManager: KeyManager, objectMapper: ObjectMapper): EncryptionService {
-        return EncryptionService(encryptionSecretService, keyManager, objectMapper)
+    fun encryptionService(encryptionSecretService: EncryptionSecretService, secretStore: SecretStore, objectMapper: ObjectMapper): EncryptionService {
+        return EncryptionService(encryptionSecretService, secretStore, objectMapper)
     }
 }