+++ libs/core/src/main/kotlin/io/stereov/singularity/hash/service/HashSecretService.kt
@@ -3,16 +3,16 @@ package io.stereov.singularity.hash.service
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.stereov.singularity.global.properties.AppProperties
 import io.stereov.singularity.global.util.Constants
-import io.stereov.singularity.secrets.core.component.KeyManager
+import io.stereov.singularity.secrets.core.component.SecretStore
 import io.stereov.singularity.secrets.core.service.SecretService
 import org.springframework.stereotype.Service
 
 @Service
 class HashSecretService(
-    keyManager: KeyManager,
+    secretStore: SecretStore,
     appProperties: AppProperties
 ) : SecretService(
-    keyManager,
+    secretStore,
     Constants.HASH_SECRET,
     "HmacSHA256",
     appProperties,