+++ libs/core/src/main/kotlin/io/stereov/singularity/secrets/core/component/SecretStore.kt
@@ -3,7 +3,7 @@ package io.stereov.singularity.secrets.core.component
 import io.stereov.singularity.secrets.core.exception.model.SecretKeyNotFoundException
 import io.stereov.singularity.secrets.core.model.Secret
 
-interface KeyManager {
+interface SecretStore {
 
     suspend fun getOrNull(key: String): Secret?
     suspend fun get(key: String): Secret {