+++ libs/core/src/main/kotlin/io/stereov/singularity/secrets/core/exception/handler/SecretsExceptionHandler.kt
@@ -2,7 +2,7 @@ package io.stereov.singularity.secrets.core.exception.handler
 
 import io.stereov.singularity.global.exception.BaseExceptionHandler
 import io.stereov.singularity.secrets.core.exception.SecretsException
-import io.stereov.singularity.secrets.core.exception.model.NoCurrentEncryptionKeyException
+import io.stereov.singularity.secrets.core.exception.model.NoCurrentKeyException
 import io.stereov.singularity.secrets.core.exception.model.SecretKeyNotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ControllerAdvice
@@ -14,7 +14,7 @@ class SecretsExceptionHandler : BaseExceptionHandler<SecretsException> {
 
     override fun getHttpStatus(ex: SecretsException) = when (ex) {
         is SecretKeyNotFoundException -> HttpStatus.INTERNAL_SERVER_ERROR
-        is NoCurrentEncryptionKeyException -> HttpStatus.INTERNAL_SERVER_ERROR
+        is NoCurrentKeyException -> HttpStatus.INTERNAL_SERVER_ERROR
         else -> HttpStatus.INTERNAL_SERVER_ERROR
     }