+++ libs/core/src/main/kotlin/io/stereov/singularity/secrets/core/exception/model/SecretStoreException.kt
@@ -2,4 +2,4 @@ package io.stereov.singularity.secrets.core.exception.model
 
 import io.stereov.singularity.secrets.core.exception.SecretsException
 
-class KeyManagerException(msg: String, cause: Throwable? = null) : SecretsException(msg, cause)
+class SecretStoreException(msg: String, cause: Throwable? = null) : SecretsException(msg, cause)