+++ libs/core/src/main/kotlin/io/stereov/singularity/secrets/core/properties/SecretStoreProperties.kt
@@ -3,8 +3,8 @@ package io.stereov.singularity.secrets.core.properties
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties("singularity.secrets")
-data class KeyManagerProperties(
-    val keyManager: KeyManagerImplementation = KeyManagerImplementation.Bitwarden,
+data class SecretStoreProperties(
+    val store: SecretStoreImplementation = SecretStoreImplementation.Bitwarden,
     val keyRotationCron: String = "0 0 4 1 1,4,7,10 *",
     val cacheExpiration: Long = 900000,
 )